Doubles tonen

+++ Android/src/com/ilmoitus/activity/ChangeDeclareActivity.java
@@ -385,18 +385,11 @@ public class ChangeDeclareActivity extends Activity{
 				new GetAttachments(activity, attachments).execute();
 				for (int i = 0; i < lines.length(); i++) {
 					JSONObject line = lines.getJSONObject(i);
-					JSONObject declarationType = (JSONObject) line
-							.get("declaration_type");
-					JSONObject declarationSubType = (JSONObject) line
-							.get("declaration_sub_type");
-					DeclarationLine decLine = new DeclarationLine(
-							line.getLong("id"), line.getString("receipt_date"),
-							new DeclarationTypes(declarationType
-									.getString("name"), declarationType
-									.getLong("id")),
-							new DeclarationSubTypes(declarationSubType
-									.getString("name"), declarationSubType
-									.getLong("id")), line.getInt("cost"));
+					JSONObject declarationType = (JSONObject) line.get("declaration_type");
+					JSONObject declarationSubType = (JSONObject) line.get("declaration_sub_type");
+					DeclarationLine decLine = new DeclarationLine(line.getLong("id"), line.getString("receipt_date"),
+							new DeclarationTypes(declarationType.getString("name"), declarationType.getLong("id")),
+							new DeclarationSubTypes(declarationSubType.getString("name"), declarationSubType.getLong("id")), line.getDouble("cost"));
 					declarationLines.add(decLine);
 				}
 				DeclarationLineAdapter decLineAdapter = new DeclarationLineAdapter(
